feat(cli): adding support for HID transport and fixing derivation path encoder

+++ sn_cli/src/subcommands/ledger/get_pub_key.rs
@@ -22,7 +22,7 @@ pub struct GetPubKeyReq {
 }
 
 impl GetPubKeyReq {
-    pub fn new(path: &[u8]) -> Self {
+    pub fn new(path: &[u32]) -> Self {
         Self {
             dpath: derivation_path(path),
         }
